Добавлен функционал (начальный) просмотра списка новостей, а так же добавление новости

+++ app/Post.php
@@ -13,16 +13,16 @@ class Post extends Model
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category()
     {
-    	return $this->hasOne(Category::class);
+    	return $this->belongsTo(Category::class);
     }
 
     public function author()
     {
-    	return $this->hasOne(User::class);
+    	return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags()
@@ -62,17 +62,23 @@ class Post extends Model
     //remove psot function
     public function remove()
     {
-        Storage::delete('uploads/' . $this->$image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage()
+    {
+        if ($this->image != null) {
+            Storage::delete('uploads/' . $this->$image);
+        }
     }
     //upload thumnail image for post function
     public function uploadImage($image)     
     {
         if ($image == null) { return; }
-
-        Storage::delete('uploads/' . $this->$image);
+        $this->removeImage();
         $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAs('uploads', $filename);
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -81,7 +87,7 @@ class Post extends Model
     {
         if ($this->image == null) 
         {
-            return '/img/no-image.png';    
+            return '/img/no_image_post.png';
         }
 
         return '/uploads/' . $this->image;
@@ -97,7 +103,7 @@ class Post extends Model
     //set tags for post functions
     public function setTags($ids)
     {
-        if ($id == null) { return; }
+        if ($ids == null) { return; }
 
         $this->tags()->sync($ids);
     }
@@ -142,6 +148,21 @@ class Post extends Model
             return $this->setStandart();
         }
         return $this->setFeatured();
+    }
+
+    public function getCategoryTitle()
+    {
+        return ($this->category != null)
+            ? $this->category->title
+            : 'Нет категории';
+    }
+
+    public function getTagsTitles()
+    {
+
+        return(!$this->tags->isEmpty())
+            ?   implode(', ', $this->tags->pluck('title')->all())
+            :   'Нет тегов';
     }
 
 }
